# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/v1",
    "paths": {
        "/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "user login",
                "parameters": [
                    {
                        "description": "login param",
                        "name": "loginParam",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/apimodel.LoginParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":\"ok\",\"code\":200,\"data\":{\"token\":\"9a22723c3589fd6922f65cbef2310b541e7721ca351f2cf4a17a9d84e6b9599e\"}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"status\":\"error\",\"code\":401,\"msg\":\"Wrong user name or password\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logout"
                ],
                "summary": "user logout",
                "parameters": [
                    {
                        "type": "string",
                        "description": "bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":\"ok\",\"code\":200,\"data\":\"admin: logged out\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"status\":\"error\",\"code\":401,\"msg\":\"Wrong token or token expires\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apimodel.LoginParam": {
            "type": "object",
            "required": [
                "password",
                "user"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "user": {
                    "type": "string",
                    "example": "admin"
                }
            }
        }
    }
}